requirements.txt
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
alembic==1.12.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.5
Werkzeug==3.0.1
cloudinary==1.36.0
python-dotenv==1.0.1
PyJWT==2.8.0
PyYAML==6.0.2
gunicorn==21.2.0
wtforms==3.1.1
flask-swagger==0.2.14
